updating miniprintf

+++ srcs/ft_diouxx.c
@@ -6,7 +6,7 @@
 /*   By: mabouce <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/17 16:50:23 by mabouce           #+#    #+#             */
-/*   Updated: 2019/03/05 18:46:15 by mabouce          ###   ########.fr       */
+/*   Updated: 2019/04/18 16:46:45 by mabouce          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ char	*ft_int(t_stru *stru)
 	long long	i;
 	char		*tmp;
 
-	i = va_arg(stru->va, long long);
+	i = va_arg(stru->va, int);
 	tmp = ft_lltoa(i);
 	return (tmp);
 }
